pyproject.toml
[build-system]  # https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]  # https://python-poetry.org/docs/pyproject/
name = "genaistudio-agi-v4"
version = "0.0.0"
description = "A Python app that reticulates splines."
authors = ["Joren Bex, Michael Bauwens, Heike Pauli, Jonas Buysse"]
readme = "README.md"
repository = "https://ucll.openproject.com/projects/genai-studio/"
package-mode = false

[tool.poetry.dependencies]  # https://python-poetry.org/docs/dependency-specification/
streamlit = "1.41.1"
poethepoet = "^0.33.0"
python = ">=3.10.9,<4.0"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.8"
langchain-mistralai = "^0.2.6"
azure-ai-inference = "^1.0.0b9"
azure-core = "^1.32.0"

[tool.poetry.group.test.dependencies]  # https://python-poetry.org/docs/master/managing-dependencies/
coverage = { extras = ["toml"], version = ">=7.4.4" }
mypy = "^1.15.0"
pre-commit = ">=3.7.0"
pytest = "^8.3.5"
pytest-mock = ">=3.14.0"
pytest-xdist = ">=3.5.0"
ruff = ">=0.5.7"

[tool.poetry.group.dev.dependencies]  # https://python-poetry.org/docs/master/managing-dependencies/
cruft = ">=2.15.0"
ipykernel = ">=6.29.4"
ipython = ">=8.18.0"
ipywidgets = ">=8.1.2"
pdoc = ">=14.4.0"

[tool.coverage.report]  # https://coverage.readthedocs.io/en/latest/config.html#report
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.run]  # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
command_line = "--module pytest"
data_file = "reports/.coverage"
source = ["src"]

[tool.coverage.xml]  # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"

[tool.mypy]  # https://mypy.readthedocs.io/en/latest/config_file.html
junit_xml = "reports/mypy.xml"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[tool.pytest.ini_options]  # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--color=yes --doctest-modules --exitfirst --failed-first --verbosity=2 --junitxml=reports/pytest.xml"
testpaths = ["src", "tests"]
xfail_strict = true

[tool.ruff]  # https://github.com/charliermarsh/ruff
fix = true
line-length = 100
src = ["src", "tests"]
target-version = "py310"

[tool.ruff.lint]
select = ["A", "ASYNC", "B", "C4", "C90", "D", "DTZ", "E", "F", "FLY", "FURB", "I", "ISC", "LOG", "N", "NPY", "PERF", "PGH", "PIE", "PL", "PT", "Q", "RET", "RUF", "RSE", "SIM", "TID", "UP", "W", "YTT"]
ignore = ["D203", "D213", "E501", "PGH002", "PGH003", "RET504", "S101", "S307"]
unfixable = ["F401", "F841"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.poe.tasks]  # https://github.com/nat-n/poethepoet

  [tool.poe.tasks.app]
  help = "Serve the app"

    [[tool.poe.tasks.app.sequence]]
    cmd = "echo 'Serving app...'"
  [tool.poe.tasks.run]
  help = "Run the Streamlit app"
  cmd = "streamlit run app.py"

[tool.poe.tasks.ruff]
cmd = "ruff check ."

[tool.poe.tasks.ruff-fix]
cmd = "ruff check . --fix ."

[tool.poe.tasks.coverage]
cmd = "coverage run -m pytest"

[tool.poe.tasks.coverage-report]
cmd = "coverage report"

[tool.poe.tasks.coverage-html]
cmd = "coverage html"

[tool.poe.tasks.mypy]
cmd = "mypy ."

[tool.poe.tasks.test]
cmd = "pytest"

  [tool.poe.tasks.docs]
  help = "Generate this app's docs"
  cmd = """
    pdoc
      --docformat $docformat
      --output-directory $outputdirectory
      src
    """

    [[tool.poe.tasks.docs.args]]
    help = "The docstring style (default: numpy)"
    name = "docformat"
    options = ["--docformat"]
    default = "numpy"

    [[tool.poe.tasks.docs.args]]
    help = "The output directory (default: docs)"
    name = "outputdirectory"
    options = ["--output-directory"]
    default = "docs"

  [tool.poe.tasks.lint]
  help = "Lint this app"

    [[tool.poe.tasks.lint.sequence]]
    cmd = """
      pre-commit run
        --all-files
        --color always
      """
